+++ src/ralph_assets/models_assets.py
@@ -346,10 +346,18 @@ class Asset(TimeTrackable, EditorTrackable, SavingUser, SoftDeletable):
 def create_asset_post_save(sender, instance, created, **kwargs):
     """When a new DC asset without a device linked to it is created, try to
     match it with an existing device or create a dummy (stock) device and
-    match with it instead.
+    match with it instead. Note: it does not apply to assets created with
+    'add part' button.
     """
-    if created and not instance.device_info.ralph_device_id:
-        instance.create_stock_device()
+    if created:
+        try:
+            ralph_device_id = instance.device_info.ralph_device_id
+        except AttributeError:
+            # asset created with 'add part'
+            pass
+        else:
+            if not ralph_device_id:
+                instance.create_stock_device()
 
 
 class DeviceInfo(TimeTrackable, SavingUser, SoftDeletable):